requirements.txt
asgiref==3.7.2
Babel==2.12.1
build==0.10.0
Django==5.0.6
django-ckeditor==6.5.1
django-cors-headers==4.3.1
django-environ==0.10.0
django-js-asset==2.0.0
django-money==3.1.0
djangorestframework==3.15.1
drf-yasg==1.21.7
inflection==0.5.1
packaging==23.1
Pillow==9.5.0
psycopg2-binary==2.9.6
py-moneyed==2.0
pyproject_hooks==1.0.0
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.6.3
uritemplate==4.1.1
